Vistas arregladas

+++ resources/views/administration/roles/roles_index.blade.php
@@ -1,10 +1,12 @@
 @extends('app')
 
 @section('content')
+    @if(Auth::user()->canSeeIf(['crear-roles']))
 
     <div>
         <a href="/administracion/roles/create" class="btn btn-lg btn-primary">Agregar nuevo rol</a>
     </div>
+    @endif
     <hr>
 
     @if($roles->count() > 0)
@@ -17,7 +19,9 @@
                 <th>
                     Permisos
                 </th>
+                @if(Auth::user()->canSeeIf(['editar-roles','eliminar-roles']))
                 <th>Acciones</th>
+                @endif
                 </thead>
                 <tbody>
                 @foreach($roles as $role)
@@ -35,9 +39,11 @@
                                 <p>No tiene permisos asignados</p>
                             @endif
                         </td>
+                        @if(Auth::user()->canSeeIf(['editar-roles','eliminar-roles']))
                         <td>
                             @if($role->slug != 'administrador-del-sistema')
                                 <ul class="list-inline">
+                                    @if(Auth::user()->canSeeIf(['editar-roles']))
                                        <li>
                                            <a
                                                    href="/administracion/roles/{{$role->slug}}/edit"
@@ -45,7 +51,10 @@
                                                    data-toggle="tooltip"
                                                    title="Editar!">
                                            </a>
-                                       </li>|
+                                       </li>
+                                    @endif
+                                    @if(Auth::user()->canSeeIf(['eliminar-roles']))
+                                        |
                                         <li>
                                             <a class="fa fa-lg fa-times"
                                                onclick="deleteConfirm('{{$role->name}}','{{$role->id}}','/administracion/roles/')"
@@ -53,9 +62,11 @@
                                                title="Eliminar!">
                                             </a>
                                         </li>
+                                        @endif
                                 </ul>
                             @endif
                         </td>
+                        @endif
                     </tr>
                 @endforeach
                 </tbody>
